Optimize code & overclock pico
The code was heavily optimized and simplified. Some of the methods used include:
- ROMC in/out commands are processed at the same time on the rising edge of WRITE. This should be acceptable behavior according to the F8 specifications. 
- The polling loop was improved
- A few common functions like writing to dbus were inlined

The RPi Pico itself was overclocked to 428 MHz @ 1.3V. The flash clock divider has been increased from 2 to 4, allowing it to operate at a max frequency of 532 MHz. Well above what can be achieved on the RPi Pico.